+++ non_catalog_apps/ghost_esp_app/src/settings_ui.c
@@ -411,8 +411,8 @@ static void settings_action_callback(void* context, uint32_t index) {
 }
 
 // submenu callback to open wifi hardware settings menu from settings actions
-#define WIFI_SETTINGS_MENU_ID   200
-#define STATUS_IDLE_MENU_ID     201
+#define WIFI_SETTINGS_MENU_ID 200
+#define STATUS_IDLE_MENU_ID   201
 static void wifi_settings_menu_callback(void* context, uint32_t index) {
     UNUSED(index);
     AppState* state = (AppState*)context;