+++ non_catalog_apps/ghost_esp_app/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.6.1
+
+- Added multiple status display animations from @tototo31
+- Added IR Dazzler menu option
+
 ## v1.6.0
 
 - Added IR menu